refactor compressed radix tree internals
